requirements.txt
aiogram~=3.2.0
requests~=2.31.0
Pillow~=10.1.0
googletrans~=4.0.0rc1
langcodes~=3.3.0